+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Fernschach-Verwaltung Changelog
 
+## Version 0.18.3 (2024-03-08)
+
+* Change: Filter in tl_fernschach_spieler "Nicht Mitglied nach 31.12.JJJJ" von -9 auf -16 Jahre erhöht
+
 ## Version 0.18.2 (2024-01-27)
 
 * Change: Drei Filter mit Namen "Geprüft" -> geändert auf "Geprüft K", "Geprüft B", "Geprüft N" (Konto, Beitrag, Nenngeld)
 
 ## Version 0.18.1 (2024-01-20)
 
-* Add: Filter in tl_fernschach_spieler -> "Mitglieder neu in 2024", der Beitritte in 2024 wiedergibt (ega ob Neu- oder Wiedereintritt)
+* Add: Filter in tl_fernschach_spieler -> "Mitglieder neu in 2024", der Beitritte in 2024 wiedergibt (egal ob Neu- oder Wiedereintritt)
 * Add: Helper-Funktion isMember zur Ermittlung einer Mitgliedschaft an einem bestimmten Datum oder in einem bestimmten Jahr
 
 ## Version 0.18.0 (2024-01-17)